chore: separate out generic worker from callback_worker
