+++ pooler/utils/callback_worker.py
@@ -1,11 +1,7 @@
 import asyncio
 import importlib
 import json
-import multiprocessing
-import resource
-import signal
 import time
-from functools import partial
 from typing import Callable
 from typing import Dict
 from typing import List
@@ -15,23 +11,13 @@ from typing import Union
 from uuid import uuid4
 
 from aio_pika import IncomingMessage
-from aio_pika.pool import Pool
-from eth_utils import keccak
-from httpx import AsyncClient
-from httpx import AsyncHTTPTransport
-from httpx import Limits
-from httpx import Timeout
 from pydantic import ValidationError
-from redis import asyncio as aioredis
-from setproctitle import setproctitle
 
 from pooler.settings.config import projects_config
 from pooler.settings.config import settings
 from pooler.utils.callback_helpers import AuditProtocolCommandsHelper
-from pooler.utils.callback_helpers import get_rabbitmq_channel
-from pooler.utils.callback_helpers import get_rabbitmq_connection
 from pooler.utils.callback_helpers import notify_on_task_failure
-from pooler.utils.default_logger import logger
+from pooler.utils.generic_worker import GenericAsyncWorker
 from pooler.utils.models.data_models import PayloadCommitAPIRequest
 from pooler.utils.models.data_models import SnapshotterIssue
 from pooler.utils.models.data_models import SnapshotterIssueSeverity
@@ -39,109 +25,23 @@ from pooler.utils.models.data_models import SourceChainDetails
 from pooler.utils.models.message_models import PowerloomCallbackProcessMessage
 from pooler.utils.models.message_models import SnapshotBase
 from pooler.utils.redis.rate_limiter import load_rate_limiter_scripts
-from pooler.utils.redis.redis_conn import RedisPoolCache
 from pooler.utils.redis.redis_keys import (
     cb_broadcast_processing_logs_zset,
 )
 from pooler.utils.redis.redis_keys import discarded_query_epochs_redis_q
 from pooler.utils.redis.redis_keys import failed_query_epochs_redis_q
-from pooler.utils.rpc import RpcHelper
 
 
-class CallbackAsyncWorker(multiprocessing.Process):
+class SnapshotAsyncWorker(GenericAsyncWorker):
 
     def __init__(self, name, **kwargs):
-        self._core_rmq_consumer: asyncio.Task
-        self._q = f'powerloom-backend-cb:{settings.namespace}:{settings.instance_id}'
-        self._rmq_routing = f'powerloom-backend-callback:{settings.namespace}:{settings.instance_id}.*'
-        self._unique_id = f'{name}-' + keccak(text=str(uuid4())).hex()[:8]
-        self._aioredis_pool = None
-        self._redis_conn: Union[None, aioredis.Redis] = None
-        self._running_callback_tasks: Dict[str, asyncio.Task] = dict()
-        super(CallbackAsyncWorker, self).__init__(name=name, **kwargs)
-        self._logger = logger.bind(module=self.name)
+        super(SnapshotAsyncWorker, self).__init__(name=name, **kwargs)
 
-        self._rate_limiting_lua_scripts = None
-        self._client = None
-        self._async_transport = None
-        self._shutdown_signal_received_count = 0
-        self._rpc_helper = None
         self._project_calculation_mapping = None
-
-    async def _shutdown_handler(self, sig, loop: asyncio.AbstractEventLoop):
-        self._shutdown_signal_received_count += 1
-        if self._shutdown_signal_received_count > 1:
-            self._logger.info(
-                (
-                    f'Received exit signal {sig.name}. Not processing as'
-                    ' shutdown sequence was already initiated...'
-                ),
-            )
-        else:
-            self._logger.info(
-                (
-                    f'Received exit signal {sig.name}. Processing shutdown'
-                    ' sequence...'
-                ),
-            )
-            # check the done or cancelled status of self._running_callback_tasks.values()
-            for u_uid, t in self._running_callback_tasks.items():
-                self._logger.debug(
-                    (
-                        'Shutdown handler: Checking result and status of'
-                        ' aio_pika consumer callback task {}'
-                    ),
-                    t.get_name(),
-                )
-                try:
-                    task_result = t.result()
-                except asyncio.CancelledError:
-                    self._logger.info(
-                        (
-                            'Shutdown handler: aio_pika consumer callback task'
-                            ' {} was cancelled'
-                        ),
-                        t.get_name(),
-                    )
-                except asyncio.InvalidStateError:
-                    self._logger.info(
-                        (
-                            'Shutdown handler: aio_pika consumer callback task'
-                            ' {} result not available yet. Still running.'
-                        ),
-                        t.get_name(),
-                    )
-                except Exception as e:
-                    self._logger.info(
-                        (
-                            'Shutdown handler: aio_pika consumer callback task'
-                            ' {} raised Exception. {}'
-                        ),
-                        t.get_name(),
-                        e,
-                    )
-                else:
-                    self._logger.info(
-                        (
-                            'Shutdown handler: aio_pika consumer callback task'
-                            ' returned with result {}'
-                        ),
-                        t.get_name(),
-                        task_result,
-                    )
-
-            tasks = [
-                t
-                for t in asyncio.all_tasks(loop)
-                if t is not asyncio.current_task(loop)
-            ]
-
-            [task.cancel() for task in tasks]
-
-            self._logger.info(f'Cancelling {len(tasks)} outstanding tasks')
-            await asyncio.gather(*tasks, return_exceptions=True)
-            loop.stop()
-            self._logger.info('Shutdown complete.')
+        self._task_types = []
+        for project_config in projects_config:
+            type_ = project_config.project_type
+            self._task_types.append(type_)
 
     @notify_on_task_failure
     async def _processor_task(self, msg_obj: PowerloomCallbackProcessMessage, task_type: str):
@@ -566,25 +466,11 @@ class CallbackAsyncWorker(multiprocessing.Process):
             json.dumps(update_state),
         )
 
-    async def _rabbitmq_consumer(self, loop):
-        self._rmq_connection_pool = Pool(get_rabbitmq_connection, max_size=5, loop=loop)
-        self._rmq_channel_pool = Pool(
-            partial(get_rabbitmq_channel, self._rmq_connection_pool), max_size=20,
-            loop=loop,
-        )
-        async with self._rmq_channel_pool.acquire() as channel:
-            await channel.set_qos(20)
-            q_obj = await channel.get_queue(
-                name=self._q,
-                ensure=False,
-            )
-            self._logger.debug(
-                f'Consuming queue {self._q} with routing key {self._rmq_routing}...',
-            )
-            await q_obj.consume(self._on_rabbitmq_message)
-
     async def _on_rabbitmq_message(self, message: IncomingMessage):
         task_type = message.routing_key.split('.')[-1]
+        if task_type not in self._task_types:
+            return
+
         self._logger.debug('task type: {}', task_type)
         await message.ack()
 
@@ -611,34 +497,6 @@ class CallbackAsyncWorker(multiprocessing.Process):
 
         asyncio.ensure_future(self._processor_task(msg_obj=msg_obj, task_type=task_type))
 
-    async def _init_redis_pool(self):
-        if self._aioredis_pool is not None:
-            return
-        self._aioredis_pool = RedisPoolCache()
-        await self._aioredis_pool.populate()
-        self._redis_conn = self._aioredis_pool._aioredis_pool
-
-    async def _init_httpx_client(self):
-        if self._client is not None:
-            return
-        self._async_transport = AsyncHTTPTransport(
-            limits=Limits(
-                max_connections=100,
-                max_keepalive_connections=50,
-                keepalive_expiry=None,
-            ),
-        )
-        self._client = AsyncClient(
-            timeout=Timeout(timeout=5.0),
-            follow_redirects=False,
-            transport=self._async_transport,
-        )
-
-    async def _init_rpc_helper(self):
-        if self._rpc_helper is not None:
-            return
-        self._rpc_helper = RpcHelper()
-
     async def _init_project_calculation_mapping(self):
         if self._project_calculation_mapping is not None:
             return
@@ -657,31 +515,3 @@ class CallbackAsyncWorker(multiprocessing.Process):
         await self._init_httpx_client()
         await self._init_rpc_helper()
         await self._init_project_calculation_mapping()
-
-    def run(self) -> None:
-        setproctitle(self._unique_id)
-        soft, hard = resource.getrlimit(resource.RLIMIT_NOFILE)
-        resource.setrlimit(
-            resource.RLIMIT_NOFILE,
-            (settings.rlimit.file_descriptors, hard),
-        )
-        # logging.config.dictConfig(config_logger_with_namespace(self.name))
-        ev_loop = asyncio.get_event_loop()
-        signals = (signal.SIGTERM, signal.SIGINT, signal.SIGQUIT)
-        for s in signals:
-            ev_loop.add_signal_handler(
-                s,
-                lambda x=s: ev_loop.create_task(
-                    self._shutdown_handler(x, ev_loop),
-                ),
-            )
-        self._logger.debug(
-            f'Starting asynchronous epoch callback worker {self._unique_id}...',
-        )
-        self._core_rmq_consumer = asyncio.ensure_future(
-            self._rabbitmq_consumer(ev_loop),
-        )
-        try:
-            ev_loop.run_forever()
-        finally:
-            ev_loop.close()
